+++ CHANGES.md
@@ -13,6 +13,20 @@
 
 ### misc
 
+- [CHANGE] GitHub Actions の ubuntu-latest を ubuntu-24.04 に変更する
+  - @voluntas
+
+## 2024.8.0
+
+- [ADD] 採用する結果の信頼スコアの最小値を指定する minimum_confidence_score を追加する
+  - Amazon Transcribe のみ有効
+  - デフォルト値: 0（信頼スコアを無視する）
+  - @Hexa
+- [ADD] 採用する結果の最小発話期間（秒）を指定する minimum_transcribed_time を追加する
+  - Amazon Transcribe のみ有効
+  - デフォルト値: 0（最小発話期間を無視する）
+  - @Hexa
+
 ## 2024.7.0
 
 - [FIX] サービスへの接続が成功してもリトライカウントがリセットされない不具合を修正する